requirements.txt
numpy
scipy
tensorflow-gpu==2.6.0
Pillow
matplotlib
jupyter
tqdm
